requirements.txt
speedtests
requests